requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.9.2
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.11=pypi_0
aiosignal=1.3.2=pypi_0
aisuite=0.1.6=pypi_0
alembic=1.14.0=pypi_0
altair=5.5.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.7.0=pypi_0
appdirs=1.4.4=pypi_0
appnope=0.1.4=pypi_0
arxiv=2.1.3=pypi_0
asgiref=3.8.1=pypi_0
asttokens=3.0.0=pypi_0
attrs=24.3.0=pypi_0
auth0-python=4.7.2=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.2.1=pypi_0
beautifulsoup4=4.12.3=pypi_0
blinker=1.9.0=pypi_0
bs4=0.0.2=pypi_0
build=1.2.2.post1=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.11.26=hca03da5_0
cachetools=5.5.0=pypi_0
certifi=2024.12.14=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.0=pypi_0
chroma-hnswlib=0.7.6=pypi_0
chromadb=0.5.23=pypi_0
click=8.1.7=pypi_0
cohere=5.13.3=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.2.2=pypi_0
crewai=0.86.0=pypi_0
crewai-tools=0.17.0=pypi_0
cryptography=43.0.3=pypi_0
dataclasses-json=0.6.7=pypi_0
debugpy=1.8.11=pypi_0
decorator=5.1.1=pypi_0
deprecated=1.2.15=pypi_0
deprecation=2.1.0=pypi_0
distro=1.9.0=pypi_0
docker=7.1.0=pypi_0
docstring-parser=0.16=pypi_0
docx2txt=0.8=pypi_0
durationpy=0.9=pypi_0
embedchain=0.1.125=pypi_0
et-xmlfile=2.0.0=pypi_0
executing=2.1.0=pypi_0
expat=2.6.4=h313beb8_0
faiss-cpu=1.9.0.post1=pypi_0
fastapi=0.115.6=pypi_0
fastavro=1.9.7=pypi_0
feedparser=6.0.11=pypi_0
filelock=3.16.1=pypi_0
flatbuffers=24.3.25=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.12.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-api-core=2.24.0=pypi_0
google-auth=2.37.0=pypi_0
google-cloud-aiplatform=1.75.0=pypi_0
google-cloud-bigquery=3.27.0=pypi_0
google-cloud-core=2.4.1=pypi_0
google-cloud-resource-manager=1.14.0=pypi_0
google-cloud-storage=2.19.0=pypi_0
google-crc32c=1.6.0=pypi_0
google-resumable-media=2.7.2=pypi_0
googleapis-common-protos=1.66.0=pypi_0
gptcache=0.1.44=pypi_0
groq=0.9.0=pypi_0
grpc-google-iam-v1=0.13.1=pypi_0
grpcio=1.68.1=pypi_0
grpcio-status=1.68.1=pypi_0
grpcio-tools=1.68.1=pypi_0
h11=0.14.0=pypi_0
h2=4.1.0=pypi_0
hpack=4.0.0=pypi_0
httpcore=1.0.7=pypi_0
httptools=0.6.4=pypi_0
httpx=0.27.2=pypi_0
httpx-sse=0.4.0=pypi_0
huggingface-hub=0.27.0=pypi_0
humanfriendly=10.0=pypi_0
hyperframe=6.0.1=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
iniconfig=2.0.0=pypi_0
instructor=1.7.0=pypi_0
ipykernel=6.29.5=pypi_0
ipython=8.30.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.4=pypi_0
jiter=0.6.1=pypi_0
json-repair=0.32.0=pypi_0
jsonpatch=1.33=pypi_0
jsonpickle=4.0.1=pypi_0
jsonpointer=3.0.0=pypi_0
jsonref=1.1.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
kubernetes=31.0.0=pypi_0
lancedb=0.17.0=pypi_0
langchain=0.3.13=pypi_0
langchain-cohere=0.3.3=pypi_0
langchain-community=0.3.13=pypi_0
langchain-core=0.3.27=pypi_0
langchain-experimental=0.3.3=pypi_0
langchain-groq=0.2.2=pypi_0
langchain-openai=0.2.14=pypi_0
langchain-text-splitters=0.3.4=pypi_0
langchainhub=0.1.21=pypi_0
langgraph=0.2.60=pypi_0
langgraph-checkpoint=2.0.9=pypi_0
langgraph-sdk=0.1.48=pypi_0
langserve=0.3.0=pypi_0
langsmith=0.1.147=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
litellm=1.55.6=pypi_0
mako=1.3.8=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
marshmallow=3.23.2=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mem0ai=0.1.36=pypi_0
mmh3=5.0.1=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.0=pypi_0
multidict=6.1.0=pypi_0
mypy-extensions=1.0.0=pypi_0
narwhals=1.19.1=pypi_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.6.0=pypi_0
networkx=3.4.2=pypi_0
nodeenv=1.9.1=pypi_0
numpy=1.26.4=pypi_0
oauthlib=3.2.2=pypi_0
onnxruntime=1.20.1=pypi_0
openai=1.58.1=pypi_0
openpyxl=3.1.5=pypi_0
openssl=3.0.15=h80987f9_0
opentelemetry-api=1.29.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.29.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.29.0=pypi_0
opentelemetry-exporter-otlp-proto-http=1.29.0=pypi_0
opentelemetry-instrumentation=0.50b0=pypi_0
opentelemetry-instrumentation-asgi=0.50b0=pypi_0
opentelemetry-instrumentation-fastapi=0.50b0=pypi_0
opentelemetry-proto=1.29.0=pypi_0
opentelemetry-sdk=1.29.0=pypi_0
opentelemetry-semantic-conventions=0.50b0=pypi_0
opentelemetry-util-http=0.50b0=pypi_0
orjson=3.10.12=pypi_0
outcome=1.3.0.post0=pypi_0
overrides=7.7.0=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
parameterized=0.9.0=pypi_0
parso=0.8.4=pypi_0
pdfminer-six=20231228=pypi_0
pdfplumber=0.11.4=pypi_0
pexpect=4.9.0=pypi_0
phidata=2.7.5=pypi_0
pillow=11.0.0=pypi_0
pip=24.2=py312hca03da5_0
platformdirs=4.3.6=pypi_0
pluggy=1.5.0=pypi_0
portalocker=2.10.1=pypi_0
posthog=3.7.4=pypi_0
prompt-toolkit=3.0.48=pypi_0
propcache=0.2.1=pypi_0
proto-plus=1.25.0=pypi_0
protobuf=5.29.2=pypi_0
psutil=6.1.1=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pyarrow=18.1.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pycparser=2.22=pypi_0
pydantic=2.10.4=pypi_0
pydantic-core=2.27.2=pypi_0
pydantic-settings=2.7.0=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.18.0=pypi_0
pyjwt=2.10.1=pypi_0
pylance=0.20.0=pypi_0
pypdf=5.1.0=pypi_0
pypdf2=3.0.1=pypi_0
pypdfium2=4.30.1=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.2.0=pypi_0
pyright=1.1.391=pypi_0
pysbd=0.3.4=pypi_0
pysocks=1.7.1=pypi_0
pytest=8.3.4=pypi_0
python=3.12.8=h99e199e_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
pytube=15.0.0=pypi_0
pytz=2024.2=pypi_0
pyvis=0.3.2=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=pypi_0
qdrant-client=1.12.1=pypi_0
readline=8.2=h1a28f6b_0
referencing=0.35.1=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=13.9.4=pypi_0
rpds-py=0.22.3=pypi_0
rsa=4.9=pypi_0
schema=0.7.7=pypi_0
selenium=4.27.1=pypi_0
setuptools=75.1.0=py312hca03da5_0
sgmllib3k=1.0.0=pypi_0
shapely=2.0.6=pypi_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.6=pypi_0
sqlalchemy=2.0.36=pypi_0
sqlite=3.45.3=h80987f9_0
sse-starlette=2.2.1=pypi_0
stack-data=0.6.3=pypi_0
starlette=0.41.3=pypi_0
streamlit=1.41.1=pypi_0
sympy=1.13.3=pypi_0
tabulate=0.9.0=pypi_0
tenacity=9.0.0=pypi_0
tiktoken=0.7.0=pypi_0
tk=8.6.14=h6ba3021_0
tokenizers=0.20.3=pypi_0
toml=0.10.2=pypi_0
tomli=2.2.1=pypi_0
tomli-w=1.1.0=pypi_0
tornado=6.4.2=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
trio=0.27.0=pypi_0
trio-websocket=0.11.1=pypi_0
typer=0.15.1=pypi_0
types-requests=2.32.0.20241016=pypi_0
typing-extensions=4.12.2=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=pypi_0
uv=0.5.10=pypi_0
uvicorn=0.34.0=pypi_0
uvloop=0.21.0=pypi_0
watchfiles=1.0.3=pypi_0
wcwidth=0.2.13=pypi_0
websocket-client=1.8.0=pypi_0
websockets=14.1=pypi_0
wheel=0.44.0=py312hca03da5_0
wikipedia=1.4.0=pypi_0
wrapt=1.17.0=pypi_0
wsproto=1.2.0=pypi_0
xz=5.4.6=h80987f9_1
yarl=1.18.3=pypi_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h18a0788_1
